+++ db/upgrade.php
@@ -52,5 +52,13 @@ function xmldb_auth_ldap_syncplus_upgrade($oldversion) {
         }
     }
 
+    if ($oldversion < 2020121003) {
+        // Normalize the memberattribute_isdn plugin config.
+        set_config('memberattribute_isdn',
+                !empty(get_config('auth_ldap_syncplus', 'memberattribute_isdn')), 'auth_ldap');
+
+        upgrade_plugin_savepoint(true, 2020121003, 'auth', 'ldap_syncplus');
+    }
+
     return true;
 }